refactor(archive): cut out to a separate file
